feat: Add real-time GT vehicle position plot
Use a parallel process to perform the plot. Note that adding camera
feed to the matplotlib window is extremely slow. Could be an
inherent matplotlib limitation or a bug in my code. Some reported
using matplotlib's FuncAnimation is faster but I couldn't get it
to work (it blocks the main thread)